Split benchmark into smaller jobs.
