+++ benchmarks/bernoulli.run.jl
@@ -3,6 +3,9 @@ include("bernoulli.model.jl")
 
 bench_res = tbenchmark("HMC(1000, 0.25, 5)", "bermodel", "berdata")
 logd = build_logd("Bernoulli Model", bench_res...)
+
+
+include("bernoulli-stan.run.jl")
 logd["stan"] = Dict("theta" => mean(theta_stan))
 logd["time_stan"] = ber_time
